requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pymongo==3.12.1
pyparsing==3.1.2
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rsa==4.9
sqlparse==0.2.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
